requirements.txt
alembic==1.8.1
anyio==3.6.1
APScheduler==3.9.1
attrs==22.1.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
decorator==4.4.2
fastapi==0.78.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.3
gunicorn==20.1.0
h11==0.13.0
idna==3.4
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
login==0.0.6
lxml==4.9.1
Mako==1.2.3
MarkupSafe==2.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.30
numpy==1.23.3
packaging==21.3
plivo==4.26.0
pluggy==1.0.0
protobuf==3.20.1
psycopg2==2.9.3
py==1.11.0
pydantic==1.10.2
PyJWT==2.5.0
pyparsing==3.0.9
pytest==7.1.3
python-dotenv==0.20.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.19.1
tomli==2.0.1
twilio==7.11.0
typing_extensions==4.3.0
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
uvicorn==0.18.2
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.8.1
